Modify title
Change the first line of the file to be a web page title, replacing the link to the page. That link will remain in the README on the master branch.

The reason for this action is that MakeCode editor does not display the images, rather only links to the images. Now the users of a MakeCode editor have the choice to view the README as a web page if they wish.